Fix decreaseKey emulation bug
This commit fixes a bug in AStar, where the check for "duplicate"
priority queue entries was never satisfied for currently unknown
reasons, but most likely having to do with the serialization process.

Changelist:
- The priority queue used by AStar now takes in an instance of the
PriorityQueueEntry class, a wrapper around a vertex label and its key in
the queue.
- The Vertex class no longer contains a pqKey field, as well as any
relevant getter or setter.
- The DeserializedEdge, DeserializedGraph and DeserializedVertex classes
have been recompiled using updated schemas to reflect the above changes.
- The serialize method of the Graph class has been updated to reflect
the above changes.
- The AStar and Path classes have been received a minor general cleanup.